app/routes/utils/ephemeris_calculator.py
# ephemeris_calculator.py

from flask import Blueprint, request, jsonify
from datetime import datetime, timedelta, timezone as dt_timezone
from pytz import timezone as pytz_timezone
from skyfield.api import wgs84
from skyfield.almanac import find_discrete, sunrise_sunset
import swisseph as swe
import numpy as np
from timezonefinder import TimezoneFinder
import math
import uuid

from app.routes.constants import PLANETARY_ORDER, DAY_RULERS, ZODIAC_SIGNS, PLANETARY_COLORS, ORDINAL_NAMES, SKYFIELD_IDS, ESSENTIAL_DIGNITIES, EXTENDED_PLANETARY_ORDER, EXTENDED_SKYFIELD_IDS
from app.routes.constants import ephemeris, ts
from app.routes.constants import NEO4J_URI, NEO4J_USER, NEO4J_PASSWORD, neo4j_driver


class EphemerisCalculator:
    def __init__(self, latitude, longitude):
        self.latitude = latitude
        self.longitude = longitude
        self.observer = wgs84.latlon(latitude, longitude)

        # Initialize timezone
        tz_finder = TimezoneFinder()
        self.timezone_name = tz_finder.timezone_at(lat=latitude, lng=longitude)
        if not self.timezone_name:
            raise ValueError("Could not determine timezone for the given location.")
        self.timezone = pytz_timezone(self.timezone_name)

        # Initialize times
        self.now_utc = datetime.now(dt_timezone.utc)
        self.now_local = self.now_utc.astimezone(self.timezone)
        self.sunrise_local, self.sunset_local = self._calculate_sun_times()


    def generate_ephemeris_dataset(self):
        """
        Combine all planetary data into one unified dataset for easy access.
        """
        ephemeris_dataset = self.calculate_planetary_positions()

        # Add planetary distances
        distances = self.calculate_planetary_distances()
        for planet, distance in distances.items():
            if planet in ephemeris_dataset:
                ephemeris_dataset[planet]["distance_au"] = distance

        # Add combustion/cazimi data
        combustion_cazimi = self.calculate_combustion_and_cazimi()
        for planet, status in combustion_cazimi.items():
            if planet in ephemeris_dataset:
                ephemeris_dataset[planet].update(status)

        # Add Moon-specific properties
        moon_data = self.calculate_moon_properties()
        if "Moon" in ephemeris_dataset:
            ephemeris_dataset["Moon"].update(moon_data)

        # Add aspects (e.g., angular distances)
        aspects = self.calculate_aspects()
        for planet, aspect_data in aspects.items():
            if planet in ephemeris_dataset:
                ephemeris_dataset[planet]["angular_distance"] = aspect_data.get("angular_distance")

        return ephemeris_dataset

    def calculate_planetary_positions(self):
        # Logic for planetary positions
        pass

    def calculate_planetary_distances(self):
        # Logic for planetary distances
        pass

    def calculate_combustion_and_cazimi(self):
        # Combustion and cazimi logic
        pass

    def calculate_aspects(self):
        # Aspects logic
        pass

    def calculate_moon_properties(self):
        # Moon-specific calculations
        pass